implement simple substitution

+++ src/parser/node.h
@@ -28,8 +28,8 @@ class NStatement : public Node {
 
 class NInteger : public NExpression {
 public:
-    long long value;
-    NInteger(long long value) : value(value) { }
+    int value;
+    NInteger(int value) : value(value) { }
 //    virtual llvm::Value* codeGen(CodeGenContext& context);
 };
